Final Update on Side Menu
Side Menu with mobile responsive
