+++ Employee-Smart-Roster-main/Employee-Smart-Roster-main/fireJetReact/src/App.css
@@ -53,23 +53,6 @@
   border-bottom-color: var(--primary-background);
 }
 
-.App-side-menu {
-  width: 25%;
-  max-width: 250px;
-  background-color: var(--side-menu-background);
-  opacity: var(--side-menu-transparency);
-  color: var(--light-color-text);
-  padding: 12px;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  gap: 10px;
-  overflow-y: auto; /* Enable scrolling if content overflows */
-  z-index: 1000;
-  height: 100%; /* Make it take full height of parent */
-  flex-shrink: 0; /* Prevent shrinking */
-}
-
 .App-secondary-text {
   font-size: var(--secondary-text);
   color: var(--primary-text);
@@ -402,11 +385,6 @@ option:checked {
 
   h3 {
     font-size: var(--text-h3-m);
-  }
-
-  .App-side-menu {
-    display: none;
-    opacity: 100%;
   }
 
   .App-desktop-responsive-table {
