+++ Employee-Smart-Roster-main/Employee-Smart-Roster-main/fireJetReact/src/components/NavBar/NavBar.css
@@ -41,6 +41,7 @@ div{
     justify-content: end;
     gap: 5px;
     justify-self: end;
+    align-items: center;
   }
 
   .navbar-brand {
