+++ Employee-Smart-Roster-main/Employee-Smart-Roster-main/fireJetReact/src/components/SideMenu/SideMenu_t.tsx
@@ -1,22 +1,181 @@
-import { useAuth } from '../../AuthContext' 
-import { USER_ROLE } from '../../controller/Variables.js'
-import SASide from './SASide'
-import BOSide from './BOSide'
-import EmpSide from './EmpSide'
+import { useAuth } from "../../AuthContext";
+import { USER_ROLE } from "../../controller/Variables";
+import { BsShopWindow } from "react-icons/bs";
+import { MdOutlineBugReport } from "react-icons/md";
+import { LuUserCog } from "react-icons/lu";
+import { GrSchedules } from "react-icons/gr";
+import { FaHeartCircleExclamation, FaUsersViewfinder } from "react-icons/fa6";
+import { FaRegBuilding } from "react-icons/fa";
+import { CiStar } from "react-icons/ci";
 
-import './menu.css'
-import '../../../public/styles/common.css'
+import Menu from "./Menu";
 
-const SideMenu_t = ({}) => {
+export const SA_Items = [
+    {
+        name: 'user',
+        label: 'USER MANAGEMENT',
+        src: <LuUserCog className="menu-icon"/>,
+        items: [
+            {
+                name: 'userManagement',
+                label: 'BUSINESS OWNER MANAGEMENT',
+                navHref: '/users-management',
+            },
+            {
+                name: 'userRegReq',
+                label: 'Registration Request Management',
+                navHref: '/registration-req-management',
+            }
+        ]
+    },
+    {
+        name: 'issue',
+        label: 'ISSUES REPORTED',
+        navHref: '/issues-reported',
+        src: <MdOutlineBugReport className="menu-icon"/>,
+    },
+    {
+        name: 'landing',
+        label: 'LANDING PAGE MANAGEMENT',
+        src: <BsShopWindow className="menu-icon"/>,
+        items: [
+            {
+                name: 'landingDemoVideo',
+                label: 'Demo Video Management',
+                navHref: '/video-management'
+            },
+            {
+                name: 'landingReview',
+                label: 'Reviews & Ratings',
+                navHref: '/review-rating'
+            },
+            {
+                name: 'landingFAQ',
+                label: 'FAQs Management',
+                navHref: '/faqs-management'
+            },
+            {
+                name: 'landingPreview',
+                label: 'Preview Landing Page',
+                navHref: '/preview-landing-page'
+            },
+        ]
+    },
+]
+
+export const BO_Items = [
+        {
+            name: 'user',
+            label: 'COMPANY',
+            src: <FaRegBuilding className="menu-icon"/>,
+            items: [
+                {
+                    name: 'companyProfile',
+                    label: 'My Company',
+                    navHref: '/company-detail',
+                },
+                {
+                    name: 'subscriptionManagement',
+                    label: 'Subscription Management',
+                    navHref: '/subscription-management',
+                }
+            ]
+        },
+        {
+            name: 'timelineManagement',
+            label: 'TIMELINE MANAGEMENT',
+            navHref: '/timeline-management',
+            src: <GrSchedules className="menu-icon"/>,
+        },
+        {
+            name: 'employee',
+            label: 'MY EMPLOYEE',
+            src: <FaUsersViewfinder className="menu-icon"/>,
+            items: [
+                {
+                    name: 'employeeManagement',
+                    label: 'Employee Management',
+                    navHref: '/users-management'
+                },
+                {
+                    name: 'attendanceRecords',
+                    label: 'Attendance Records',
+                    navHref: '/attendance-records-management'
+                },
+                {
+                    name: 'leaveManagement',
+                    label: 'Leave Management',
+                    navHref: '/mc-management'
+                },
+            ]
+        },
+        {
+            name: 'reportIssues',
+            label: 'REPORT ISSUES',
+            navHref: '/report-issues',
+            src: <MdOutlineBugReport className="menu-icon"/>,
+        },
+        {
+            name: 'reviwRating',
+            label: 'REVIEW & RATING MANAGEMENT',
+            navHref: '/review-n-rating-management',
+            src: <CiStar className="menu-icon"/>,
+        },
+]
+
+export const EMP_Items = [
+    {
+        name: 'timelineManagement',
+        label: 'TIMELINE MANAGEMENT',
+        src: <GrSchedules className="menu-icon"/>,
+        items: [
+            {
+                name: 'mySchedule',
+                label: 'My Schedules',
+                navHref: '/timeline-management',
+            },
+            {
+            name: 'attendanceRecord',
+            label: 'Attendance Record',
+            navHref: '/view-attendance-record',
+            },
+            {
+            name: 'swapTime',
+            label: 'Swap Time Management',
+            navHref: '/swap-time-management',
+            },
+        ]
+    },
+    {
+        name: 'leaveManagement',
+        label: 'LEAVE MANAGEMENT',
+        navHref: '/my-leave-management',
+        src: <FaHeartCircleExclamation className="menu-icon"/>,
+    },
+    {
+        name: 'reportIssues',
+        label: 'REPORT ISSUES',
+        navHref: '/report-issues',
+        src: <MdOutlineBugReport className="menu-icon"/>,
+    },
+]
+
+const SideMenu_t = () => {
     const { user } = useAuth();
-
-    return(
+    
+    return (
         <>
-        {user?.role === USER_ROLE[0] && <SASide />}
-        {user?.role === USER_ROLE[1] && <BOSide />}
-        {user?.role === USER_ROLE[2] && <EmpSide />}
+        {user?.role === USER_ROLE[0] && (
+            <Menu menuItems={SA_Items} responsive="desktop" />
+        )}
+        {user?.role === USER_ROLE[1] && (
+            <Menu menuItems={BO_Items} responsive="desktop" />
+        )}
+        {user?.role === USER_ROLE[2] && (
+            <Menu menuItems={EMP_Items} responsive="desktop" />
+        )}
         </>
-    )
-}
+    );
+};
 
 export default SideMenu_t;
